Fix Windows warning
 ../pcsc_scan.c: In function 'stress':
../pcsc_scan.c:431:39: warning: format '%ld' expects argument of type 'long int', but argument 3 has type 'size_t' {aka 'long long unsigned int'} [-Wformat=]
  431 |                 printf("%sFAPDU n°: %ld\n", cpl, count);
      |                                     ~~^          ~~~~~
      |                                       |          |
      |                                       long int   size_t {aka long long unsigned int}
      |                                     %lld

+++ pcsc_scan.c
@@ -425,10 +425,10 @@ static LONG stress(SCARDCONTEXT hContext2, const char *readerName)
 
 
 #define COUNT 100
-	size_t count;
+	int count;
 	for (count=0; count<COUNT; count++)
 	{
-		printf("%sFAPDU n°: %ld\n", cpl, count);
+		printf("%sFAPDU n°: %d\n", cpl, count);
 		dwSendLength = sizeof(pbSendBuffer);
 		dwRecvLength = sizeof(pbRecvBuffer);
 		rv = SCardTransmit(hCard, pioSendPci, pbSendBuffer, dwSendLength,
